@@ -36,65 +36,38 @@
           class="actions"
           @click="$event.stopPropagation()"
         >
-          <DownloadIcon
-            title="Export"
-            @click="exportSavedPattern(pattern.name)"
-          />
+          <div class="icon">
+            <i class="material-icons" title="Export" @click="exportSavedKnitPattern(pattern.name)">download</i>
+          </div>
 
-          <FontDownloadIcon
-            title="Export Knit"
-            @click="exportSavedKnitPattern(pattern.name)"
-          />
 
-          <DeleteIcon
-            danger
-            title="Delete"
-            @click="deletePattern(pattern.name)"
-          />
+          <div class="icon">
+            <i class="material-icons" title="Delete" @click="deletePattern(pattern.name)">delete</i>
+          </div>
         </div>
       </div>
     </div>
   </div>
 </template>
 
-<script>
-import { mapActions, mapGetters } from 'vuex';
-import DeleteIcon from 'vue-material-design-icons/Delete.vue';
-import DownloadIcon from 'vue-material-design-icons/Download.vue';
-import FontDownloadIcon from 'vue-material-design-icons/FountainPen.vue';
+<script setup lang="ts">
+import { onMounted, useTemplateRef } from 'vue';
+import { storeToRefs } from 'pinia';
+import { usePersistanceStore } from '@/store/persistance';
 import PanelButton from '../inputs/PanelButton.vue';
 
-export default {
-  components: {
-    PanelButton,
-    DeleteIcon,
-    DownloadIcon,
-    FontDownloadIcon,
-  },
-  data: () => ({
-    uploadReady: true,
-  }),
-  computed: {
-    ...mapGetters(['savedPatterns']),
-  },
-  created() {
-    this.loadIndex();
-  },
-  methods: {
-    ...mapActions([
-      'loadIndex',
-      'loadPattern',
-      'importPattern',
-      'deletePattern',
-      'updateSettings',
-      'exportSavedKnitPattern',
-    ]),
-    async runImport(e) {
-      await this.importPattern(e);
-      this.$refs.importInput.type = 'text';
-      this.$refs.importInput.type = 'file';
-    },
-  },
+const persistanceStore = usePersistanceStore();
+const { savedPatterns } = storeToRefs(persistanceStore);
+const { loadIndex, loadPattern, importPattern, deletePattern, exportSavedKnitPattern } = persistanceStore;
+
+const importInputEl = useTemplateRef('importInput')
+
+onMounted(() => loadIndex());
+
+const runImport = async (e: Event) => {
+  await importPattern(e as InputEvent);
+  importInputEl.value!.type = 'text';
+  importInputEl.value!.type = 'file';
 };
 </script>
 
@@ -103,21 +76,26 @@ export default {
   display: flex;
   justify-content: space-between;
 }
+
 .header {
   margin-bottom: 10px;
   display: flex;
   justify-content: space-between;
   align-items: center;
 }
+
 .import-button {
   width: 100px;
 }
+
 .pattern-count {
   font-size: 20px;
 }
+
 .pattern-count .number {
   font-weight: bold;
 }
+
 .saved-patterns {
   height: 311px;
   max-height: 311px;
@@ -125,6 +103,7 @@ export default {
   border-radius: 5px;
   overflow-y: scroll;
 }
+
 .saved-pattern-row {
   border-bottom: 1px solid grey;
   padding: 0px 10px;
@@ -134,20 +113,25 @@ export default {
   position: relative;
   cursor: pointer;
 }
+
 .saved-pattern-row:hover {
   background-color: #eee;
 }
+
 .saved-pattern-row:last-child:nth-child(n+7) {
   border-bottom: none;
 }
-.saved-pattern-row > span {
+
+.saved-pattern-row>span {
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
 }
-.saved-pattern-row:hover > span {
+
+.saved-pattern-row:hover>span {
   width: 210px;
 }
+
 .actions {
   position: absolute;
   height: 100%;
@@ -157,10 +141,15 @@ export default {
   opacity: 0;
   transition: opacity 0.2s;
 }
+
 .saved-pattern-row:hover .actions {
   opacity: 1;
+  display: flex;
+  align-items: center;
 }
-.actions > * {
+
+.actions>.icon {
+  height: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -168,11 +157,13 @@ export default {
   user-select: none;
   transition: all 0.2s;
 }
-.actions > *:hover {
+
+.actions>*:hover {
   background-color: var(--safe-color);
   color: white;
 }
-.actions > *[danger]:hover {
+
+.actions>*[danger]:hover {
   background-color: var(--danger-color);
   color: white;
 }
