@@ -9,7 +9,7 @@
 
     <div
       class="container"
-      :style="{ maxHeight: open ? maxHeightOpen : 0 }"
+      :style="<CSSProperties>{ maxHeight: open ? maxHeightOpen : 0 }"
     >
       <div
         ref="content"
@@ -21,25 +21,30 @@
   </section>
 </template>
 
-<script>
-export default {
-  props: {
-    name: { type: String, required: true },
-    open: { type: Boolean, default: false },
-  },
-  data: () => ({
-    maxHeightOpen: '0',
-  }),
-  mounted() {
-    this.maxHeightOpen = `${this.$refs.content.getBoundingClientRect().height}px`;
-  },
-};
+<script setup lang="ts">
+import { onMounted, ref, useTemplateRef, type CSSProperties } from 'vue';
+
+defineProps<{
+  name: string,
+  open?: boolean,
+}>();
+
+defineEmits<{ 'update:open': [open: boolean] }>();
+
+const contentEl = useTemplateRef('content');
+
+const maxHeightOpen = ref('0');
+
+onMounted(() => {
+  maxHeightOpen.value = `${contentEl.value!.getBoundingClientRect().height}px`;
+});
 </script>
 
 <style scoped>
 section {
   border-top: 2px solid grey;
 }
+
 .heading {
   display: flex;
   justify-content: center;
@@ -49,14 +54,17 @@ section {
   user-select: none;
   cursor: pointer;
 }
-.open > .heading {
+
+.open>.heading {
   font-weight: bold;
 }
+
 .container {
   max-height: 0;
   overflow: hidden;
   transition: max-height 0.2s;
 }
+
 .content {
   padding: 15px 20px;
 }
