@@ -7,12 +7,9 @@
   </button>
 </template>
 
-<script>
-export default {
-  props: {
-    danger: { type: Boolean, default: false },
-  },
-};
+<script setup lang="ts">
+defineProps<{ danger?: boolean }>();
+defineEmits<{ click: [$event: MouseEvent] }>();
 </script>
 
 <style scoped>
