@@ -1,17 +1,17 @@
-/* eslint-disable class-methods-use-this */
-class FileHelper {
-  static async importRaw(e) {
+export default class FileHelper {
+  static async importRaw(e: InputEvent) {
     return new Promise((resolve) => {
-      const file = e.target.files[0];
+      const file = (e.target as HTMLInputElement)?.files?.[0];
+      if (!file) return;
       const reader = new FileReader();
-      reader.onload = (evt) => {
-        resolve(evt.target.result);
+      reader.onload = (evt: ProgressEvent<FileReader>) => {
+        resolve(evt.target?.result);
       };
       reader.readAsText(file);
     });
   }
 
-  static download(filename, text) {
+  static download(filename: string, text: string) {
     const element = document.createElement('a');
     element.setAttribute('href', `data:text/plain;charset=utf-8,${encodeURIComponent(text)}`);
     element.setAttribute('download', filename);
@@ -21,5 +21,3 @@ class FileHelper {
     document.body.removeChild(element);
   }
 }
-
-export default FileHelper;
