@@ -1,7 +1,10 @@
+import type { GridSquares } from '../models/grid';
+import type { KnitPattern } from '../models/knit';
+import type { CompressedPattern } from '../models/pattern';
 import PatternHelper from './pattern-helper';
 
-class PersistanceHelper {
-  static colorString(color) {
+export default class PersistanceHelper {
+  static colorString(color: string) {
     switch (color) {
       case '#ffffff': return 'White';
       case '#000000': return 'Black';
@@ -9,7 +12,7 @@ class PersistanceHelper {
     }
   }
 
-  static createKnitString(name, reducedPatternWithSettings) {
+  static createKnitString(name: string, reducedPatternWithSettings: KnitPattern) {
     let text = `-------- ${name} --------\r\n\r\n`;
 
     reducedPatternWithSettings.forEach((row, index) => {
@@ -24,11 +27,11 @@ class PersistanceHelper {
     return text;
   }
 
-  static compressPattern(sortedPattern) {
-    const colorDict = {};
+  static compressPattern(sortedPattern: GridSquares): CompressedPattern {
+    const colorDict: Record<string, string> = {};
     let patternString = '';
 
-    let previousColorAlias;
+    let previousColorAlias: string = '';
     let colorCount = 0;
 
     sortedPattern.forEach((point) => {
@@ -58,17 +61,17 @@ class PersistanceHelper {
     };
   }
 
-  static decompressPattern({ dim, col, pat }) {
+  static decompressPattern({ dim, col, pat }: CompressedPattern) {
     const splitPat = pat.split(',');
     const blankPattern = PatternHelper.createFilledPattern(dim, '#ffffff');
 
-    let left;
-    let color;
+    let left: number | null;
+    let color: string;
 
     const pattern = blankPattern.map((point) => {
       if (!left) {
-        const split = splitPat.shift().split('-');
-        color = Object.keys(col).find((x) => col[x] === split[0]);
+        const split = splitPat.shift()!.split('-');
+        color = Object.keys(col).find((x) => col[x] === split[0])!;
         left = Number(split[1]);
       }
       const newPoint = { ...point, color };
@@ -79,5 +82,3 @@ class PersistanceHelper {
     return pattern;
   }
 }
-
-export default PersistanceHelper;
